dwarf-reader: Bug 31377 - Fix the IR for zero length arrays
Before C99, the zero-length array GCC extension described at
https://gcc.gnu.org/onlinedocs/gcc-4.4.2/gcc/Zero-Length.html would be
represented in DWARF by a DW_TAG_subrange_type DIE with a DW_AT_count
attribute set to zero.

Since the C99 standard, standardized flexible arrays are now the way
to go.  They are represented in DWARF by a DW_TAG_subrange_type that
has no DW_AT_lower_bound, no DW_AT_upper_bound and no DW_AT_count
attribute.

Both zero-length and flexible arrays represent in essence the same
thing.  So they should be represented by the same libabigail IR.
Instead, the libabigail IR represents them slightly differently.  The
former has the upper_bound of the subrange set to -1 while the latter
has it set to zero.  Both have the 'is_infinite' property set to true,
meaning that the subrange has an unknown (flexible) number of
elements.

Because the upper_bound of the two sub-ranges are different, the
equals overload for array_type_def::subrange_type returns false,
(wrongly) indicating that the two sub-ranges are different.

In practice, that leads to abidiff emitting spurious changes between
the two types below which should normally represent the same thing:

    int foo[0]; // gcc zero-length extension
    int foo[]; // c99 flexible array

This patch makes the DWARF reader set the upper_bound of the sub-range
type to zero for zero-length arrays, just like what is done for
flexible arrays.  The "is-infinite" property of the sub-range type is
set to true to flag the type as being for a flexible array.

This fixes the issue described above.

	* src/abg-dwarf-reader.cc (build_subrange_type): In the absence of
	DW_AT_upper_bound and DW_AT_lower_bound, make DW_AT_count being
	zero the same as if no DW_AT_count is present.
	* src/abg-reader.cc (build_subrange_type): Add a comment about how
	things were in the past (before this patch) and document the fact
	that we want to stay compatible with ABIXML from before this
	patch.
	* tests/data/test-abidiff-exit/PR31377/flex/include/rte_pipeline.h:
	New header file input for this test.
	* tests/data/test-abidiff-exit/PR31377/flex/lib64/librte_pipeline.so.24.1:
	New binary input for this test.
	* tests/data/test-abidiff-exit/PR31377/reference/include/rte_pipeline.h:
	New header file input for this test.
	* tests/data/test-abidiff-exit/PR31377/reference/lib64/librte_pipeline.so.24.0:
	New binary input for this test.
	* tests/data/test-abidiff-exit/PR31377/test-PR31377-report-1.txt:
	New reference test output.
	* tests/data/Makefile.am: Add the
	* tests/data/test-diff-dwarf-abixml/PR25409-librte_bus_dpaa.so.20.0.abi:
	Adjust.
	* tests/data/test-read-dwarf/PR25007-sdhci.ko.abi: Likewise.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Likewise.
	* tests/test-abidiff-exit.cc

Signed-off-by: Dodji Seketeli <dodji@redhat.com>