PR25409-librte_bus_dpaa.so.20.0.abi: Update to 2.3
The PR25409-librte_bus_dpaa.so.20.0.abi test input file was modified
at least because of the patch to fix zero-length arrays IR for
https://sourceware.org/bugzilla/show_bug.cgi?id=31377.

That modification came after the introduction of the pointer-to-member
type feature, which bumped the ABIXML version to 2.3.  So,
PR25409-librte_bus_dpaa.so.20.0.abi should see its ABIXML version
change to 2.3 too.  This patch makes that adjustment.

The reason why the patch to fix zero-length arrays IR hasn't bumped
the ABIXML version to 2.3 has to do with easing the potential
back-porting into the libabigail-2.4-branch.

	* tests/data/test-diff-dwarf-abixml/PR25409-librte_bus_dpaa.so.20.0.abi:
	Bump ABIXML version to 2.3.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>