ir: decl-only classes don't equal fully defined classes under ODR
Until now, when the One Definition Rule is in effect (when analyzing
C++, for instance) we assume that a decl-only class equals all the
fully-defined classes of the same name.  Now that we are using type
hashing to determine type equality during type canonicalization, we
cannot keep that assumption, because the hash of a decl-only class
does not necessarily equals the hash of its definition.  Actually,
decl-only classes don't have hashes.  So this patch removes that
assumption.

Note that even with this patch, "make check" still yields 3 failures
that are still XFAILed in the test suite.

	* src/abg-ir.cc (equals): In language supporting the ODR (just
	like in all other languages), consider that a decl-only class is
	different from its defined class counterpart.
	* tests/data/test-diff-dwarf/test28-vtable-changes-report-0.txt:
	Adjust.
	* tests/data/test-diff-filter/test41-report-0.txt: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>