Bug 31513 - abidiff considers data members moved to base class as harmful
Consider the 'struct type' below and its use in the file
example-v0.cc:

    //----> example-v0.cc: <------
    struct type
    {
      int m0;
    };

    int
    use(type& t)
    {
      return t.m0;
    }
    //---->8<-----

Now, consider a second version of that program where the data member
of the 'struct type' is moved into a base class:

    //----> example-v1.cc: <------
    struct base
    {
      int m0;
    };

    struct type : public base
    {
    };

    int
    use(type& t)
    {
      return t.m0;
    }
    //---->8<------

The first and second version of 'struct type' are ABI compatible
because neither the size of 'struct type' nor the offset of the data
member m0 has changed.

But then look at what abidiff says:

    $ abidiff example-v0.o example-v1.o || echo "exit code: $?"
    Functions changes summary: 0 Removed, 1 Changed, 0 Added function
    Variables changes summary: 0 Removed, 0 Changed, 0 Added variable

    1 function with some indirect sub-type change:

      [C] 'function int use(type&)' at example-v0.cc:7:1 has some indirect sub-type changes:
	parameter 1 of type 'type&' has sub-type changes:
	  in referenced type 'struct type' at example-v1.cc:6:1:
	    type size hasn't changed
	    1 base class insertion:
	      struct base at example-v1.cc:1:1
	    1 data member deletion:
	      'int m0', at offset 0 (in bits) at example-v0.cc:3:1

    exit code: 4
    $

abidiff considers that the mere addition of a base class constitutes a
potential ABI incompatibility.

Also, abidiff fails to see that although the data member m0 was
removed from 'struct type' itself, it was moved into its new base
type.  In practice, the data member m0 is still a member of 'struct
type', by definition of class type inheritance.

Thus, the exit code of abidiff is ABIDIFF_ABI_CHANGE which value is 4.

This patch fixes this problem.

First of all, here is what abidiff now says with this patch:

    $ abidiff example-v0.o example-v1.o && echo "return code: $?"
    Functions changes summary: 0 Removed, 0 Changed (1 filtered out), 0 Added function
    Variables changes summary: 0 Removed, 0 Changed, 0 Added variable

    return code: 0
    $

In other words, abidiff now recognizes the changes as being ABI
compatible and filters them out by default.

Furthermore, here are the changes that abidiff sees when asked to
avoid filtering these (now considered harmless) changes:

    $ abidiff --harmless example-v0.o example-v1.o && echo "return code: $?"
    Functions changes summary: 0 Removed, 1 Changed, 0 Added function
    Variables changes summary: 0 Removed, 0 Changed, 0 Added variable

    1 function with some indirect sub-type change:

      [C] 'function int use(type&)' at example-v0.cc:7:1 has some indirect sub-type changes:
	parameter 1 of type 'type&' has sub-type changes:
	  in referenced type 'struct type' at example-v1.cc:6:1:
	    type size hasn't changed
	    1 base class insertion:
	      struct base at example-v1.cc:1:1

    $

We can see that abidiff no longer considers that the data member m0
was deleted.

At its core, this patch teaches
class_diff::ensure_lookup_tables_populated to detect data members
being moved to base classes.  That detection prevents the code base to
consider those data members as being removed from the class.

Now that the moved data members are not considered as removed, the
patch goes further to consider data member moves to base classes that
are not accompanied with any size or offset change as being harmless,
and thus, to be filtered out by default.

To that end, the patch teaches categorize_harmful_diff_node that an
added base classes is not necessarily a harmful change if it is not
accompanied with a size or offset change.  Rather, an added base
classes (or any change to a class) not accompanied with a size or
offset change is now categorized into
HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY.  Note that the previous
HARMLESS_UNION_CHANGE_CATEGORY was renamed into
HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY.

Incidentally, a removal of a base class or any change accompanied with
an offset change is always categorized as harmful change.

The patch also adds several non-regression tests.

	* include/abg-comparison.h (typedef string_decl_base_sptr_map):
	Move this typedef to include/abg-fwd.h:
	(HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY): Inside the enum
	diff_category, rename the enumerator
	HARMLESS_UNION_CHANGE_CATEGORY into this.
	(EVERYTHING_CATEGORY): Adjust this enum to refer to
	HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY, no more HARMLESS_UNION_CHANGE_CATEGORY
	* include/abg-fwd.h (typedef string_decl_base_sptr_map): Move this
	typedef here from abg-comparison.h
	(collect_non_anonymous_data_members): Declare new function.
	* src/abg-comp-filter.cc (has_offset_changes)
	(type_has_offset_changes, has_offset_changes)
	(has_subtype_changes, class_diff_has_only_harmless_changes):
	Define new static functions.
	(base_classes_removed): Rename base_classes_added_or_removed into
	this and update comment.
	(categorize_harmless_diff_node): Changes that were categorized as
	HARMLESS_UNION_CHANGE_CATEGORY is now categorized as
	HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY.  Also, a diff that
	satisfies either union_diff_has_harmless_changes or
	class_diff_has_only_harmless_changes is categorized as
	HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY.
	(categorize_harmful_diff_node): Adjust to invoke
	base_classes_removed instead of base_classes_added_or_removed.  An
	added base class doesn't necessarily qualifies as a harmful change
	anymore.  Rather, a diff node that carries an offset change as
	detected by type_has_offset_changes is categorized as
	SIZE_OR_OFFSET_CHANGE_CATEGORY.
	* src/abg-comparison.cc (get_default_harmless_categories_bitmap):
	Adjust to use
	abigail::comparison::HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY
	rather than abigail::comparison::HARMLESS_UNION_CHANGE_CATEGORY.
	(operator<<(ostream& o, diff_category c)): Likewise.
	(class_diff::ensure_lookup_tables_populated): Detect data members
	that are moved into base classes to avoid considering those data
	members as being removed from the class.
	* src/abg-default-reporter.cc (default_reporter::report): Adjust
	to use HARMLESS_UNION_OR_CLASS_CHANGE_CATEGORY rather than
	HARMLESS_UNION_CHANGE_CATEGORY.
	* src/abg-ir.cc (collect_non_anonymous_data_members): Define
	function overloads.
	* tests/data/test-abidiff-exit/PR31513/non-regr/libtest[1-4]-v{0,1}.so:
	New binary test input files.
	* tests/data/test-abidiff-exit/PR31513/non-regr/report[1-4.txt:
	New reference output files.
	* tests/data/test-abidiff-exit/PR31513/non-regr/test[1-4]-v{0,1}.cc:
	Source code of the binary input files above.
	* tests/data/test-abidiff-exit/PR31513/reported/PR31513-reported-report-{1,2}.txt:
	New reference output files.
	* tests/data/test-abidiff-exit/PR31513/reported/libloremipsum_gcc{7,11}.so:
	New binary test input files.
	* tests/data/test-abidiff-exit/PR31513/reported/libstdcpp.suppr:
	New test suppression file.
	* tests/data/Makefile.am: Add new test files above to source
	distribution.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the new input
	test files to this test harness.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-0.txt:
	Adjust.
	* tests/data/test-diff-pkg/tbb-4.1-9.20130314.fc22.x86_64--tbb-4.3-3.20141204.fc23.x86_64-report-1.txt:
	Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>