ctf-reader: During re-initialization, only clear canonicalize-able types
The patch
https://sourceware.org/git/?p=libabigail.git;a=commit;h=666555665cc4fcfc8ae19661c489822e0df00ae3
introduced a mistake: It frees the CTF archive object during the
re-initialization of the reader between each binary (kernel or
module).  The archive contains CTF type information for a kernel and
all its modules.  So it should be kept around until all those binaries
are analyzed.  Instead, that patch frees the CTF archive object after
handling each binary.  Oops.

This patch fixes that problem by free-ing the CTF archive only when
the reader itself is freed, presumably, after analyzing all binaries.

Similarly, the type map maintained by the reader contains types for
the kernel and all its modules.  So it should not be freed at
re-initialization time.  Rather, what should be freed is only the
types that are to be canonicalized at the end of processing of a given
binary.  So the patch makes the reader maintain types to be
canonicalized in a vector and clears that vector at each
re-initialization.

At re-initialization time, the patch also avoids resetting the corpus
group that the current reader feeds.

	* src/abg-ctf-reader.cc (reader::types_to_canonicalize): Add data
	member.
	(reader::add_type): Add the new type to the vector of types to be
	canonicalized.  Update comment.
	(reader::canonicalize_all_types): Now that the reader maintains
	the vector of types to be canonicalized, just pass that vector to
	canonicalize_types.  reader::types_map contains all the types that
	have been created in all the binaries processed by this reader so
	far.  Many of these types have already been canonicalized at the
	end of the analysis of binaries that have already been processed.
	Only the types created during the processing of the current binary
	have not yet be canonicalized and reader::types_to_canonicalize is
	where they are maintained.  So reader::types_map should be left
	alone by this function.
	(reader::initialize): Do not close the CTF archive here.  Do not
	clear reader::{types_map, corpus_group} either.
	(reader::read_corpus): Use nullptr, not NULL.  Make sure to not
	re-open an archive that has already been opened.
	(reader::~reader): Mark a closed archive.  This is more cosmetic
	than anything else.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>